Revert "new src file with new components made using MUI, old src file renamed…"
