--- client/src-Old/main.jsx
@@ -1,32 +1,0 @@
-import ReactDOM from 'react-dom/client'
-import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import 'bootstrap/dist/css/bootstrap.min.css'
-
-import App from './App.jsx'
-import SearchRecipes from './pages/SearchRecipes'
-import SavedRecipes from './pages/SavedRecipes'
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-    errorElement: <h1 className='display-2'>Wrong page!</h1>,
-    children: [
-      {
-        index: true,
-        element: <SearchRecipes />
-      }, {
-        path: '/saved',
-        element: <SavedRecipes />
-      },
-      // {
-      //   path: '/create',
-      //   element: <CreateRecipes />
-      // },
-    ]
-  }
-])
-
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <RouterProvider router={router} />
-)
