+++ client/src/App.css
@@ -1,55 +1,61 @@
-:root{
-  --radish: #FF0A57;
-  --radishDark: #B2003A;
-  --leafy: #77C54D;
-  --leafyDark: #478D2D;
+/* #root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
 }
 
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: 'Roboto';
+.logo {
+  height: 6em;
+  padding: 1.5em;
+  will-change: filter;
+  transition: filter 300ms;
 }
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
+}
+.logo.react:hover {
+  filter: drop-shadow(0 0 2em #61dafbaa);
+}
+
+@keyframes logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  a:nth-of-type(2) .logo {
+    animation: logo-spin infinite 20s linear;
+  }
+}
+
+.card {
+  padding: 2em;
+}
+
+.read-the-docs {
+  color: #888;
+} */
 
 body {
-  /* background-color: var(--leafyDark); */
-  background-image: url(https://images.unsplash.com/photo-1606787366850-de6330128bfc?q=80&w=1000&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxleHBsb3JlLWZlZWR8OHx8fGVufDB8fHx8fA%3D%3D);
+  font-family: 'Rubik', sans-serif;
 }
 
-img {
-  width:100%;
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-family: 'Rubik Mono One', sans-serif;
 }
 
-.cardImage {
-  width:100%;
+.small {
+  font-weight: 300;
+  font-style: italic;
 }
 
-#logo {
-  width: 150px;
-}
-
-.recipeImage {
-  align-self: center;
-}
-
-.Navbar {
-  background-color: var(--radishDark);
-  color: var(--leafy);
-  margin-bottom: 10px;
-}
-
-#Searching {
-  display:flex;
-  gap: 4px;
-}
-
-footer {
-  margin-top: 10px;
-  padding: 10px;
-  /* text-align: 'center'; */
-  /* align-items: center; */
-  justify-content: center;
-  background-color: var(--radishDark);
-  width:100vw
-}
