+++ client/src/App.jsx
@@ -1,32 +1,50 @@
-import {BrowserRouter, Routes, Route, Navigate} from 'react-router-dom'
+import './App.css';
+import {
+  ApolloClient,
+  InMemoryCache,
+  ApolloProvider,
+  createHttpLink,
+} from '@apollo/client';
+import { setContext } from '@apollo/client/link/context';
+import { Outlet } from 'react-router-dom';
 
-import './App.css'
-import HomePage from './components/pages/Home'
-import Nav from './components/Navbar'
-import Login from './components/pages/Login2'
-import SignUp from './components/pages/SignUp'
-import Footer from './components/Footer'
-import Recipe from './components/pages/Recipe'
-import CheckboxList from './components/pages/list'
-// import ErrorPage from './components/pages/ErrorPage'
+import Navbar from './components/Navbar';
+import Footer from './components/Footer';
 
-export default function App() {
+// Construct our main GraphQL API endpoint
+const httpLink = createHttpLink({
+  uri: '/graphql',
+});
 
+// Construct request middleware that will attach the JWT token to every request as an `authorization` header
+const authLink = setContext((_, { headers }) => {
+  // get the authentication token from local storage if it exists
+  const token = localStorage.getItem('id_token');
+  // return the headers to the context so httpLink can read them
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : '',
+    },
+  };
+});
+
+const client = new ApolloClient({
+  // Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+});
+
+function App() {
   return (
+    <ApolloProvider client={client}>
     <>
-      <BrowserRouter>
-        <Nav/>
-        <Routes>
-          <Route path='/' element={<HomePage/>}/>
-          <Route path='/Login' element={<Login/>}/>
-          <Route path='/SignUp' element={<SignUp/>}/>
-          <Route path='/Recipes/:id' element={<Recipe/>}/>
-          <Route path='/checkbox' element={<CheckboxList/>}/>
-          {/* <Route path='/Error' element={<ErrorPage/>}/> */}
-          {/* <Route path="*" element={<Navigate to="/Error" replace />} /> */}
-        </Routes>
-        <Footer/>
-      </BrowserRouter>
+      <Navbar />
+      <Outlet />
+      <Footer />
     </>
-  )
+    </ApolloProvider>
+  );
 }
+
+export default App;
