+++ client/src/components/Footer.jsx
@@ -1,33 +1,23 @@
-import { Container, Typography, Link, Box, Breadcrumbs } from '@mui/material'
+import { useLocation, useNavigate } from 'react-router-dom';
 
-export default function Footer() {
-        
-    return(<footer>
-        <Container sx={{
-            textAlign: 'center', display:'flex', flexDirection:'column', alignItems:'center'
-        }}>
-                <Typography variant='h6' component='h3'>
-                Created by 
-                </Typography>
-            <Typography variant='body2' component='p'>
-                    <Breadcrumbs aria-label="breadcrumb" separator=' - '>
-                        <Link href="https://github.com/Gabe-Velasquez" target="_blank" rel="noopener noreferrer" color='inherit' underline='hover'>
-                            Gabe Velasquez
-                        </Link>
+const Footer = () => {
+  const location = useLocation();
+  const navigate = useNavigate();
+  return (
+    <footer className="w-100 mt-auto text-dark p-4">
+      <div className="container text-center mb-5">
+        {location.pathname !== '/' && (
+          <button
+            className="btn btn-dark mb-3"
+            onClick={() => navigate(-1)}
+          >
+            &larr; Go Back
+          </button>
+        )}
+        <h4>&copy; {new Date().getFullYear()} - What the Fork</h4>
+      </div>
+    </footer>
+  );
+};
 
-                        <Link href="https://github.com/lukehevey" target="_blank" rel="noopener noreferrer" color='inherit' underline='hover'>
-                            Luke Hevey
-                        </Link>
-
-                        <Link href="https://github.com/Achurale" target="_blank" rel="noopener noreferrer" color='inherit' underline='hover'>
-                            Chuyi Lee
-                        </Link>
-
-                        <Link href="https://github.com/Hexteezy" target="_blank" rel="noopener noreferrer" color='inherit' underline='hover'>
-                            Raul Valverde
-                        </Link>
-                    </Breadcrumbs>
-            </Typography>
-        </Container>
-    </footer>)
-}
+export default Footer;
