--- client/src/components/pages/ErrorPage.jsx
@@ -1,11 +1,0 @@
-import Box from '@mui/material'
-
-export default function Woops() {
-    return(
-        <>
-            <Box>
-                This page does not exist
-            </Box>
-        </>
-    )
-}
